// main/print/print-labels.generated.ts
export const PRINT_LABEL_LANGUAGES = [
  'en',
  'es',
  'de',
  'tr',
  'th',
  'fil',
  'fr',
  'it',
  'pt',
  'ru',
  'fa',
  'ar',
  'ur',
  'ja',
  'zh',
  'zh-tw',
  'ko',
  'id',
  'nl',
  'hi',
  'bn',
  'sq',
  'vi',
  'ne',
] as const;